+++ assetes/schema.js
@@ -1,25 +1,11 @@
 'use strict';
 
-const ANY_ERROR_RESPONSE = {
-  type: 'object',
-  properties: {
-    error: {
-      type: 'object',
-      properties: {
-        message: { type: 'string' }
-      }
-    }
-  }
-};
+const {
+  ANY_ERROR_RESPONSE,
+  VALIDATION_ERROR_RESPONSE,
+  INTERNAL_SERVER_ERROR_RESPONCE,
+} = require('./util/errorHelper.js');
 
-const VALIDATION_ERROR_RESPONSE = {
-  type: 'object',
-  properties: {
-    error: { type: 'string' },
-    message: { type: 'string' },
-    statusCode: { type: 'number' }
-  }
-};
 module.exports = {
   REGISTOR_SCHEMA: {
     schema: {
@@ -51,7 +37,8 @@ module.exports = {
             }
           },
           '4xx': ANY_ERROR_RESPONSE,
-          '400': VALIDATION_ERROR_RESPONSE
+          '400': VALIDATION_ERROR_RESPONSE,
+          '500': INTERNAL_SERVER_ERROR_RESPONCE,
         }
       }
     }
@@ -81,7 +68,8 @@ module.exports = {
           }
         },
         '4xx': ANY_ERROR_RESPONSE,
-        '400': VALIDATION_ERROR_RESPONSE
+        '400': VALIDATION_ERROR_RESPONSE,
+        '500': NTERNAL_SERVER_ERROR_RESPONCE,
       }
     }
   },
@@ -102,7 +90,8 @@ module.exports = {
           }
         },
         '4xx': ANY_ERROR_RESPONSE,
-        '400': VALIDATION_ERROR_RESPONSE
+        '400': VALIDATION_ERROR_RESPONSE,
+        '500': NTERNAL_SERVER_ERROR_RESPONCE,
       }
     }
   },
@@ -134,7 +123,8 @@ module.exports = {
             }
           },
           '4xx': ANY_ERROR_RESPONSE,
-          '400': VALIDATION_ERROR_RESPONSE
+          '400': VALIDATION_ERROR_RESPONSE,
+          '500': NTERNAL_SERVER_ERROR_RESPONCE,
         }
       }
     }
@@ -154,7 +144,8 @@ module.exports = {
           }
         },
         '4xx': ANY_ERROR_RESPONSE,
-        '400': VALIDATION_ERROR_RESPONSE
+        '400': VALIDATION_ERROR_RESPONSE,
+        '500': NTERNAL_SERVER_ERROR_RESPONCE,
       }
     }
   }
